Improve callback support for use with Puzzles
Add support for selectively removing callbacks based on their client
data pointer. This allows the same callback to be used for multiple
instances of the same client, and to only remove those entries
associated with a specific inctance.

Change the order of callback execution so that it is possible for a
recurring callback to delete itself from the list.