Add menus_build_menu() and menus_build_entry()
To support standardising the building of menus by manual means, add
two functions to allocate lay out blocks of memory.

For now, these allocate memory using malloc(). It should be assumed
that at some stage, and API update along that lines of that in the
dataxfer library will be required to allow the use of other heap
allocation regimes.