Pass filetypes into Data Transfer negotiations
Message_DataSave needs to have a filetype specified: without this,
recipients are free to reject the request based on the type as sent.

SaveAs has been sending -1, which is "untyped". This probably works for
saves to the Filer, but can be rejected by Edit and will also fail when
saving files of a specific type direct to their target application.

Update the API so that the filetype is passed to
saveas_create_dialogue() when creating a dialogue instance. This will
then be used as the data transfer protocol is initiated.

This is a breaking change to the SaveAs API, and all clients will need
to update called to saveas_create_dialogue() in order to link to the
new version of the library.